task-19.3 "fix after review"

+++ src/main/java/com/crud/tasks/service/DbService.java
@@ -1,5 +1,6 @@
 package com.crud.tasks.service;
 
+import com.crud.tasks.controller.TaskNotFoundException;
 import com.crud.tasks.domain.Task;
 import com.crud.tasks.repository.TaskRepository;
 import lombok.RequiredArgsConstructor;
@@ -26,7 +27,11 @@ public class DbService {
          return repository.save(task);
     }
 
-    public void deleteTask(Long id) {
-        repository.deleteById(id);
+    public void deleteTask(Long id) throws TaskNotFoundException {
+        if (repository.existsById(id)) {
+            repository.deleteById(id);
+        } else {
+            throw new TaskNotFoundException();
+        }
     }
 }
